Added new game button and logic behind
